Extract infer_stmt() into a separate file so that we can reuse it
